+++ src/pyspartn/spartntypes_get.py
@@ -645,9 +645,16 @@ SPARTN_PAYLOADS_GET = {
         ),
     },
     # ********************************************************************
-    # EAS-DYN
+    # EAS-DYN TODO not yet tested - no available test data source
     # ********************************************************************
-    "SPARTN-1X-EAS-DYN": {},  # TODO
+    "SPARTN-1X-EAS-DYN": {
+        "SF084": "Customer key ID",  # plain text
+        "SF085": "Dynamic key encryption type",  # plain text
+        "SF086": "Week of applicability",
+        "SF085": "Payload encryption type",
+        "SF087": "Dynamic key length",
+        "SF088": "Dynamic key",  # variable length
+    },
     # ********************************************************************
     # EAS-GRP deprecated
     # ********************************************************************