+++ RELEASE_NOTES.md
@@ -1,5 +1,12 @@
 # pyspartn Release Notes
 
+### RELEASE 0.3.3-beta
+
+ENHANCEMENTS:
+
+1. Improved handling of half-day rollovers - thanks to @jonathanmuller for contribution.
+1. Update SPARTN-1X-BPAC and SPARTN-1X-EAS-DYN handling (not tested).
+
 ### RELEASE 0.3.2-beta
 
 ENHANCEMENTS: